+++ releasenotes.md
@@ -4,6 +4,7 @@
 - set default postProcessor on init
 - redone build process with grunt
 - drop in replacement for jquery each, extend, ajax
+- setting fallbackLng to false will stop loading and looking it up
 
 ### v1.5.0
 - pass options to sync._fetchOne, use options for fetching